+++ source/benoit/benoit/app/dump.rs
@@ -21,19 +21,42 @@
 	If not, see <https://www.gnu.org/licenses/>.
 */
 
+use crate::benoit::ImageFormat;
 use crate::benoit::app::App;
 
+extern crate png;
 extern crate webp;
 
-use std::fs::write;
-use webp::Encoder;
+use std::fs::{File, write};
+use std::io::BufWriter;
 
 impl App {
 	pub fn dump(&self, path: String, image: &[u8], canvas_width: u32) {
-		let encoder = Encoder::from_rgb(&image[..], canvas_width, canvas_width);
-
-		let data = encoder.encode_lossless();
-
-		write(path, &*data).expect("unable to write image");
+		match self.image_format {
+			ImageFormat::Png  => dump_png( &path, image, canvas_width),
+			ImageFormat::Webp => dump_webp(&path, image, canvas_width),
+		}
 	}
+}
+
+fn dump_png(path: &String, image: &[u8], canvas_width: u32) {
+	let file        = File::create(path).expect("unable to create file");
+	let file_buffer = BufWriter::new(file);
+
+	let mut encoder = png::Encoder::new(file_buffer, canvas_width, canvas_width);
+	encoder.set_color(png::ColorType::Rgb);
+	encoder.set_depth(png::BitDepth::Eight);
+	encoder.set_compression(png::Compression::Fast);
+	encoder.set_srgb(png::SrgbRenderingIntent::Perceptual);
+
+	let mut writer = encoder.write_header().expect("unable to write image");
+	writer.write_image_data(image).expect("unable to write image");
+}
+
+fn dump_webp(path: &String, image: &[u8], canvas_width: u32) {
+	let encoder = webp::Encoder::from_rgb(&image[..], canvas_width, canvas_width);
+
+	let data = encoder.encode_lossless();
+
+	write(path, &*data).expect("unable to write image");
 }